+++ src/OrthogonalSphericalShellGrids.jl
@@ -28,6 +28,7 @@ include("grid_utils.jl")
 include("zipper_boundary_condition.jl")
 include("generate_tripolar_coordinates.jl")
 include("tripolar_grid.jl")
+include("grid_extensions.jl")
 include("split_explicit_free_surface.jl")
 include("with_halo.jl")
